# backend/laravel-api/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Article Membership API",
        "description": "API for managing articles, subscriptions, and user access",
        "contact": {
            "name": "API Support",
            "email": "admin@example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/plans": {
            "post": {
                "tags": [
                    "Plans"
                ],
                "summary": "Create a new plan",
                "operationId": "6ef4d9ff7065512ba4bec5f4a43b9645",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "type",
                                    "daily_article_limit",
                                    "daily_video_limit"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Pro Reader"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Access to all premium content"
                                    },
                                    "type": {
                                        "type": "string",
                                        "enum": [
                                            "PRO_READER",
                                            "PLUS_READER",
                                            "FREE"
                                        ],
                                        "example": "PRO_READER"
                                    },
                                    "daily_article_limit": {
                                        "type": "integer",
                                        "example": 10
                                    },
                                    "daily_video_limit": {
                                        "type": "integer",
                                        "example": 5
                                    },
                                    "is_active": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Plan created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Plan"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Unauthorized"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/plans/{plan}": {
            "get": {
                "tags": [
                    "Plans"
                ],
                "summary": "Get plan details",
                "operationId": "3486b5e45c3c602dd98d148563059829",
                "parameters": [
                    {
                        "name": "plan",
                        "in": "path",
                        "description": "Plan ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Plan details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Plan"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Plan not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Plans"
                ],
                "summary": "Update a plan",
                "operationId": "34e9dd3ce421fcee3a598214af2f72ef",
                "parameters": [
                    {
                        "name": "plan",
                        "in": "path",
                        "description": "Plan ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Updated Pro Reader"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Updated description"
                                    },
                                    "type": {
                                        "type": "string",
                                        "enum": [
                                            "PRO_READER",
                                            "PLUS_READER",
                                            "FREE"
                                        ]
                                    },
                                    "daily_article_limit": {
                                        "type": "integer",
                                        "example": 15
                                    },
                                    "daily_video_limit": {
                                        "type": "integer",
                                        "example": 8
                                    },
                                    "is_active": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Plan updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Plan"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Plan not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Plans"
                ],
                "summary": "Deactivate a plan",
                "operationId": "a5c9ff2507ba1608c5dd0e0ef70c091f",
                "parameters": [
                    {
                        "name": "plan",
                        "in": "path",
                        "description": "Plan ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Plan deactivated successfully"
                    },
                    "403": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Plan not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Plan": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Pro Reader"
                    },
                    "description": {
                        "type": "string",
                        "example": "Access to all premium content"
                    },
                    "type": {
                        "type": "string",
                        "enum": [
                            "PRO_READER",
                            "PLUS_READER",
                            "FREE"
                        ],
                        "example": "PRO_READER"
                    },
                    "daily_article_limit": {
                        "type": "integer",
                        "example": 10
                    },
                    "daily_video_limit": {
                        "type": "integer",
                        "example": 5
                    },
                    "is_active": {
                        "type": "boolean",
                        "example": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}